+++ CHANGELOG.rst
@@ -7,6 +7,7 @@ Ics.py changelog
 **************
 
  - Drop support for Python 2.7 and 3.3.
+ - requirements: Lock arrow version to >0.4.2
 
 **************
 0.4